Implementacion de Kruskal

+++ src/dominio/implementaciones/Edge.java
@@ -1,7 +1,7 @@
 package dominio.implementaciones;
 
 public class Edge {
-    int data; // peso del borde
+    double data; // peso del borde
     int from; // nodo
     int to; // nodo
 }
